transport: Fix race condition when dereffing transport.
Backport of revision 5173.

The pjsip_transport_dec_ref function currently decrements the reference
count of a transport, locks the transport manager, and then does an
additional check to determine if the idle transport timer should be
started.

In an environment where another thread is handling I/O it is possible
for the other thread to receive notification that the connection for
the transport has closed and to initiate transport destruction. If
transport destruction occurs before the thread invoking
pjsip_transport_dec_ref is able to acquire the transport manager lock
it will eventually end up accessing a transport that has been destroyed.
This is worsened if the transport is reused from a caching pool and
reinitialized for a new connection. The thread invoking
pjsip_transport_dec_ref may start another idle timer causing the same
timer to be scheduled twice in the timer heap.

This change checks to determine if the transport is still registered
when the transport manager lock is finally acquired before starting
the idle timer.

ASTERISK-25230
https://trac.pjsip.org/repos/ticket/1883