sip_transport_tcp/tls: Set factory on transports created from accept
The ability to re-use tcp and tls transports when a factory is
specified now depends on transport->factory being set which is a new field
in 2.5.  This was being set only on new outgoing sockets not on
incoming sockets.  The result was that a client REGISTER created a new
socket but without the factory set, the next outgoing request to the
client, OPTIONS, INVITE, etc, would attempt to create another socket
which the client would refuse.

This patch sets the factory on transports created as a result of an
accept.