sip_msg: Prevent crash on header without vptr
Occasionally a header with no vptr gets into the list of header.  This
causes a crash when printing them.  We still need to figure out how
the header got there but this patch at least prevents the crash by checking
for a non-NULL vptr before attempting to call its print function.