pjmedia/sdp.c: Crash with an invalid SDP media format description
Backport of a security issue from 2.7.2 (see ASTERISK-27582 for more info)

The media format parsing algorithm failed to catch invalid values. This could
cause users to crash if given an SDP with a invalid media format description.

When parsing the media format description this patch now properly parses the
value and returns an error status if it can't successfully parse/convert the
value.

NOTE: This patch also backported the function 'pj_strtoul3' since the fix for
the issue required it.