pjmedia/sdp.c: Crash with an invalid SDP fmtp attribute
Backport of a security issue from 2.7.2 (see ASTERISK-27583 for more info)

The fmtp retrieval function failed to catch invalid fmtp attributes.
Because of this a user would crash if given an SDP with an invalid fmtp
attribute.

When retrieving the format this patch now makes sure the fmtp attribute is
available. If not available it now returns an error status.